+++ ports/ios/VCPort/BiometricStore.swift
@@ -65,4 +65,12 @@ enum BiometricStore {
         ]
         SecItemDelete(query as CFDictionary)
     }
+
+    static func deleteAll() {
+        let query: [String: Any] = [
+            kSecClass as String: kSecClassGenericPassword,
+            kSecAttrService as String: service
+        ]
+        SecItemDelete(query as CFDictionary)
+    }
 }